change all uint to int

+++ src/c_lib/lib/method.c
@@ -21,8 +21,8 @@ void MRS_free_event(MRS_event *the_event) {
 }
 
 /** Free the memory/buffer allocation for the MRS dimensions and events within. **/
-void MRS_free_dimension(MRS_dimension *dimensions, unsigned int n) {
-  unsigned int dim, evt;
+void MRS_free_dimension(MRS_dimension *dimensions, int n) {
+  int dim, evt;
   for (dim = 0; dim < n; dim++) {
     if (DEBUG) printf("post execution dimension %d cleanup\n", dim);
     for (evt = 0; evt < dimensions[dim].n_events; evt++) {
@@ -143,13 +143,13 @@ static inline void create_plans_for_events_in_dimension(
     double coordinates_offset, int n_events, double *fraction, double *duration,
     unsigned char *is_spectral, double *rotor_frequency_in_Hz,
     double *rotor_angle_in_rad, double *magnetic_flux_density_in_T,
-    unsigned int number_of_sidebands) {
+    int number_of_sidebands) {
   int i;
   dim->count = count;
   dim->coordinates_offset = coordinates_offset;
   dim->increment = increment;
   dim->n_events = n_events;
-  dim->events = (MRS_event *)malloc(n_events * sizeof(MRS_event));
+  dim->events = (MRS_event *)malloc((size_t)n_events * sizeof(MRS_event));
 
   dim->inverse_increment = 1.0 / increment;
   dim->normalize_offset = 0.5 - (coordinates_offset * dim->inverse_increment);
@@ -161,12 +161,12 @@ static inline void create_plans_for_events_in_dimension(
   cblas_dscal(number_of_sidebands, dim->inverse_increment, plan->vr_freq, 1);
 
   for (i = 0; i < n_events; i++) {
-    dim->events[i].event_freq_amplitude = malloc_complex128(plan->size);
+    dim->events[i].event_freq_amplitude = malloc_complex128((size_t)plan->size);
     vm_double_ones(plan->size * 2, (double *)dim->events[i].event_freq_amplitude);
     cblas_dscal(plan->size, 0.0, (double *)dim->events[i].event_freq_amplitude + 1, 2);
 
     // if (*rotor_frequency_in_Hz != 0.0 && *rotor_frequency_in_Hz != 1.0e12) {
-    //   dim->events[i].freq_amplitude = malloc_double(plan->size);
+    //   dim->events[i].freq_amplitude = malloc_double((size_t)plan->size);
     //   vm_double_ones(plan->size, dim->events[i].freq_amplitude);
     // }
     MRS_set_event(&(dim->events[i]), *fraction++, *duration++, *is_spectral++,
@@ -180,10 +180,12 @@ static inline void create_plans_for_events_in_dimension(
 
   /* buffer to hold the local frequencies and frequency offset. The buffer is useful
    * when the rotor angle is off magic angle (54.735 deg). */
-  dim->local_frequency = malloc_double(scheme->n_gamma * scheme->total_orientations);
-  dim->local_phase = malloc_double(scheme->n_gamma * scheme->total_orientations);
-  dim->freq_offset = malloc_double(scheme->octant_orientations);
-  dim->freq_amplitude = malloc_double(plan->size);
+  dim->local_frequency =
+      malloc_double((size_t)(scheme->n_gamma * scheme->total_orientations));
+  dim->local_phase =
+      malloc_double((size_t)(scheme->n_gamma * scheme->total_orientations));
+  dim->freq_offset = malloc_double((size_t)scheme->octant_orientations);
+  dim->freq_amplitude = malloc_double((size_t)plan->size);
 }
 
 /**
@@ -194,10 +196,10 @@ MRS_dimension *MRS_create_dimensions(
     MRS_averaging_scheme *scheme, int *count, double *coordinates_offset,
     double *increment, double *fractions, double *durations, unsigned char *is_spectral,
     double *magnetic_flux_density_in_T, double *rotor_frequency_in_Hz,
-    double *rotor_angle_in_rad, int *n_events, unsigned int n_dim,
-    unsigned int *number_of_sidebands) {
-  unsigned int i;
-  MRS_dimension *dimension = (MRS_dimension *)malloc(n_dim * sizeof(MRS_dimension));
+    double *rotor_angle_in_rad, int *n_events, int n_dim, int *number_of_sidebands) {
+  int i;
+  MRS_dimension *dimension =
+      (MRS_dimension *)malloc((size_t)n_dim * sizeof(MRS_dimension));
 
   for (i = 0; i < n_dim; i++) {
     create_plans_for_events_in_dimension(
